Move AST types out of `symbol_table`
