+++ crates/fortitude_linter/src/rules/correctness/intent.rs
@@ -1,7 +1,8 @@
 use crate::AstRule;
 use crate::ast::FortitudeNode;
+use crate::ast::types::AttributeKind;
 use crate::settings::{CheckSettings, FortranStandard};
-use crate::symbol_table::{AttributeKind, SymbolTables};
+use crate::symbol_table::SymbolTables;
 use crate::traits::TextRanged;
 use itertools::Itertools;
 use ruff_diagnostics::{Diagnostic, Violation};
