+++ crates/fortitude_linter/src/rules/modernisation/out_of_line_attribute.rs
@@ -1,12 +1,10 @@
 use crate::ast::FortitudeNode;
+use crate::ast::types::{ParameterStatement, Variable, get_name_node_of_declarator};
 use crate::fix::edits::{
     add_attribute_to_var_decl, remove_from_comma_sep_stmt, remove_variable_decl,
 };
-use crate::symbol_table::{
-    ParameterStatement, SymbolTables, Variable, get_name_node_of_declarator,
-};
 use crate::traits::{HasNode, TextRanged};
-use crate::{AstRule, FromAstNode};
+use crate::{AstRule, FromAstNode, SymbolTables};
 
 use anyhow::{Context, Result};
 use itertools::Itertools;
